+++ src/main/scala/com/parsley/orm/compile/InstanceToSQL.scala
@@ -1,0 +1,7 @@
+package com.parsley.orm.compile
+
+object InstanceToSQL {
+    
+    
+    
+}
