+++ src/main/scala/com/parsley/orm/compile/SQLToInstance.scala
@@ -1,0 +1,5 @@
+package com.parsley.orm.compile
+
+object SQLToInstance {
+
+}
