+++ src/test/scala/Test.scala
@@ -45,7 +45,7 @@ case class Book(name: String, age: Int)
     create(books)
     val student = Student("反射魔典", 514)
 //    students.insert(student)
-    val list = students.queryRelation(books)(student)
+    val list = students.queryRelation[Book](student)
     list.foreach(println)
 }
